Sample update
git-svn-id: http://primefaces.googlecode.com/svn/examples/trunk/showcase@4566 5a70e874-a5cb-11dd-8af2-493fb0bb9f7c

+++ src/main/java/org/primefaces/examples/view/RatingController.java
@@ -16,7 +16,7 @@ public class RatingController {
 	private double rating4 = 3;
 	
 	public void handleRate(RateEvent rateEvent) {
-		FacesMessage message = new FacesMessage(FacesMessage.SEVERITY_INFO, "Rate Event", "You rated:" + (int) rateEvent.getRating());
+		FacesMessage message = new FacesMessage(FacesMessage.SEVERITY_INFO, "Rate Event", "You rated:" + ((Double) rateEvent.getRating()).intValue());
 
 		FacesContext.getCurrentInstance().addMessage(null, message);
 	}
